Enable trivial device capability test
